pyproject.toml
[project]
name = "solutionbank"
version = "0.1.0"
description = "SolutionBank Backend"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "boto3>=1.40.39",
    "cryptography>=46.0.1",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-storages>=1.14.6",
    "django-unfold>=0.66.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "google-auth>=2.40.3",
    "gunicorn>=23.0.0",
    "pillow>=11.0.0",
    "psycopg[binary]>=3.2.10",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "twilio>=9.8.8",
    "uvicorn[standard]>=0.37.0",
    "whitenoise>=6.11.0",
]

[tool.pyright]
typeCheckingMode = "off"
reportMissingImports = false
reportMissingTypeStubs = false
reportGeneralTypeIssues = false
reportOptionalMemberAccess = false
reportOptionalSubscript = false
reportOptionalOperand = false
reportOptionalCall = false
reportOptionalIterable = false
reportOptionalContextManager = false
reportUntypedFunctionDecorator = false
reportUntypedClassDecorator = false
reportUntypedBaseClass = false
reportUntypedNamedTuple = false
reportPrivateUsage = false
reportConstantRedefinition = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportInconsistentConstructor = false
reportOverlappingOverload = false
reportMissingSuperCall = false
reportUninitializedInstanceVariable = false
reportInvalidStringEscapeSequence = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportMissingParameterType = false
reportMissingTypeArgument = false
reportInvalidTypeVarUse = false
reportCallInDefaultInitializer = false
reportUnnecessaryIsInstance = false
reportUnnecessaryCast = false
reportAssertAlwaysTrue = false
reportSelfClsParameterName = false
reportImplicitStringConcatenation = false

include = ["apps", "config", "manage.py", "main.py"]
exclude = [
    "**/__pycache__",
    "**/migrations",
    ".venv",
    "venv",
    ".git"
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "coverage>=7.6.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = [
    "migrations",
    "__pycache__",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B904",  # Allow raising exceptions without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.coverage.run]
source = ["apps", "config"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/test_*.py",
    "*/admin.py",
    "config/*",
    ".venv/*",
    "venv/*",
    "manage.py",
    "main.py",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
